% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assignColors}
\alias{assignColors}
\title{Assign global colors to items}
\usage{
assignColors(items, shuffle = FALSE)
}
\arguments{
\item{items}{Vector of item names to assign colors to}

\item{shuffle}{Whether to shuffle the colors, default is FALSE}
}
\value{
A named vector of colors, with names matching the input items
}
\description{
This function assigns colors from the global_colors palette to a vector of items.
Colors will be recycled if there are more items than colors.
}
