% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateRelativeChanges}
\alias{calculateRelativeChanges}
\title{Calculate Relative Changes Between Groups}
\usage{
calculateRelativeChanges(df, group)
}
\arguments{
\item{df}{Data frame of contributions}

\item{group}{Vector of two group names}
}
\value{
Data frame with relative contribution values
}
\description{
This function calculates relative changes between two groups.
}
\keyword{internal}
