% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateSignificance}
\alias{calculateSignificance}
\title{Calculate Significance Between Groups}
\usage{
calculateSignificance(x, group)
}
\arguments{
\item{x}{Matrix of values}

\item{group}{Vector of group names}
}
\value{
p-value from comparison
}
\description{
This function calculates significance of differences between two groups.
}
\keyword{internal}
