% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{continuousColors}
\alias{continuousColors}
\title{Create a continuous color palette}
\usage{
continuousColors(low = "darkblue", mid = "white", high = "red", n = 100)
}
\arguments{
\item{low}{Color for low values, default is "darkblue"}

\item{mid}{Color for middle values, default is "white"}

\item{high}{Color for high values, default is "red"}

\item{n}{Number of colors in the palette, default is 100}
}
\value{
A vector of colors forming a continuous palette
}
\description{
This function creates a continuous color palette for use in heatmaps
and other visualizations with continuous data.
}
