% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{createStandardizedMatrix}
\alias{createStandardizedMatrix}
\title{Create Standardized Matrix With Consistent Dimensions}
\usage{
createStandardizedMatrix(mat, all_cells)
}
\arguments{
\item{mat}{Original matrix}

\item{all_cells}{Vector of all cell types to include}
}
\value{
A standardized matrix with all cells
}
\description{
Helper function to create a standardized matrix with all cell types.
}
\keyword{internal}
