% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractNet}
\alias{extractNet}
\title{Extract Network From CellChat Object}
\usage{
extractNet(object, signaling = NULL, slot.name = "netP", thresh = 0.05)
}
\arguments{
\item{object}{A CellChat object}

\item{signaling}{Signaling pathway name}

\item{slot.name}{Slot name containing the network information, default is "netP"}

\item{thresh}{Significance threshold, default is 0.05}
}
\value{
A matrix of network connections
}
\description{
This function extracts network data from a CellChat object for specific signaling pathways.
}
