% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterGroups}
\alias{filterGroups}
\title{Filter Groups From Data Frame}
\usage{
filterGroups(x, select, pair.name.all)
}
\arguments{
\item{x}{Data frame}

\item{select}{Vector of group names to select}
}
\value{
Filtered data frame
}
\description{
This function filters a data frame by group and removes empty rows.
}
\keyword{internal}
