% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/utils.R
\name{getCommonPathways}
\alias{getCommonPathways}
\title{Get Common Pathways Across CellChat Objects}
\usage{
getCommonPathways(object.list, indices = NULL)

getCommonPathways(object.list, indices = NULL)
}
\arguments{
\item{object.list}{List of CellChat objects}

\item{indices}{Vector of indices to consider}
}
\value{
Vector of common pathway names

Vector of common pathway names
}
\description{
This function extracts the signaling pathways that are common across multiple CellChat objects.

This function extracts the signaling pathways that are common across multiple CellChat objects.
}
