% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getUnionPathways}
\alias{getUnionPathways}
\title{Get union of pathways across CellChat objects}
\usage{
getUnionPathways(object.list, indices = NULL)
}
\arguments{
\item{object.list}{List of CellChat objects}

\item{indices}{Vector of indices to consider}
}
\value{
Vector of all unique pathway names
}
\description{
This function extracts all unique pathways present in any of the CellChat objects.
}
