% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{global_colors}
\alias{global_colors}
\title{Global color palette for CellDiff visualizations}
\format{
An object of class \code{character} of length 57.
}
\usage{
global_colors
}
\description{
A consistent color palette used across all CellDiff visualization functions
for cell types and other categorical variables.
}
\keyword{datasets}
