% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCell.R
\name{print.CellDiffAnalysis}
\alias{print.CellDiffAnalysis}
\title{Print method for CellDiffAnalysis}
\usage{
\method{print}{CellDiffAnalysis}(x, ...)
}
\arguments{
\item{x}{A CellDiffAnalysis object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for CellDiffAnalysis
}
