% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateCellChatObjects}
\alias{validateCellChatObjects}
\title{Check for CellChat objects and validate them}
\usage{
validateCellChatObjects(object.list, comparison = NULL)
}
\arguments{
\item{object.list}{List of CellChat objects}

\item{comparison}{Vector of indices to compare}
}
\value{
TRUE if all objects are valid, throws an error otherwise
}
\description{
This function checks if the provided objects are valid CellChat objects
with all required slots and data structures.
}
\keyword{internal}
