% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_heatmap.R
\name{view_heatmap}
\alias{view_heatmap}
\title{View ComplexHeatmap safely in RStudio Plots pane}
\usage{
view_heatmap(heatmap_obj, width = 2400, height = 1200, res = 150)
}
\arguments{
\item{heatmap_obj}{A ComplexHeatmap object or result from heatDiffM}

\item{width}{Width of the plot in pixels (default: 2400)}

\item{height}{Height of the plot in pixels (default: 1200)}

\item{res}{Resolution in DPI (default: 150)}
}
\value{
NULL (displays plot as side effect)
}
\description{
This function displays a ComplexHeatmap object in RStudio's Plots pane
without crashing. It works by rendering to PNG first, then displaying
the image.
}
\examples{
# After creating a heatmap:
hm <- heatDiffM(cellchat, comparison_method = "custom_pairs", ...)

# View it safely in RStudio:
view_heatmap(hm)

# Or with custom size:
view_heatmap(hm, width = 3000, height = 1500)

}
