% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/utils.R
\name{getCommonCellTypes}
\alias{getCommonCellTypes}
\title{Get Common Cell Types Across CellChat Objects}
\usage{
getCommonCellTypes(object.list, indices = NULL)

getCommonCellTypes(object.list, indices = NULL)
}
\arguments{
\item{object.list}{List of CellChat objects}

\item{indices}{Vector of indices to consider}
}
\value{
Vector of common cell type names

Vector of common cell type names
}
\description{
This function extracts the cell types that are common across multiple CellChat objects.

This function extracts the cell types that are common across multiple CellChat objects.
}
\examples{
\dontrun{
# Load example data
data(cellchatlist)

# Get common cell types across all objects
common_cells <- getCommonCellTypes(cellchatlist)

# Get common cell types for specific conditions
common_cells <- getCommonCellTypes(cellchatlist, indices = c(1, 2))
}

}
