% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCommonCellTypes}
\alias{getCommonCellTypes}
\title{Get common cell types across CellChat objects}
\usage{
getCommonCellTypes(object.list, indices = NULL)
}
\arguments{
\item{object.list}{List of CellChat objects}

\item{indices}{Vector of indices to consider}
}
\value{
Vector of common cell type names
}
\description{
This function extracts the cell types that are common across multiple CellChat objects.
}
